Support IPv6; remove 'save' and 'dump' features
Add IPv6 support with an ip6tables stanza in the configuration.

Each stanza now permits multiple definitions, althoug this is really
only useful if all but the first stanza in a given family are set as
non-destructive (option destroy 0) as otherwise it will clear each
table in the restore-file.

Multiple restore-file source types are now permitted as well, with
supported source formats including gz, raw, and cmd.

Remove the 'save' and 'dump' features, which won't necessarily work well
when the recently added 'destroy' is set to 0, for non-destructive
ruleset loading. It's not obvious which savefile to use now that
multiple rulesets can be loaded for each address family, and the easiest
is to get rid of this feature completely.