Change rule format parsing options to simplify use
Now the ruleset is always declared with the 'rules' option, and the
format selects the form the rules are in. The default is now 'raw' for
plaintext, with options gz, xz, and cmd supported. Other formats could
be extended if needed (eg: bzip2, or other exotic forms.)

Preserve the legacy-behavior of accepting the 'save' file, although with
the change to the 'raw' format by default, at minimum that would need to
be added to preserve backwards-compat support against earlier verisons.